House Price Prediction
After thorough research, I was able to find out different ways for feature engineering and feature selection for proper preparation of the dataset. I applied the various kinds of regression models as well in which the lasso model gave the best results where it was able to minimize the root mean squared error to 0.13. Various data preprocessing techniques have been applied on the model as well which include outlier detection, feature transformation to fix the skewness of data, data imputation and many more.